Finalized compute_all function
Changed compute_scores to compute_all and added in it the calculation of the word frequencies. This was done instead of creating an unnecessary function compute_frequencies. Most importantly, it also implies less computing time since it avoids looping over the words twice!
The test for the function was modified in accordance.