renamed module, new module, game plan
First, the initialize.py module was renamed probabilities.py as this made more sense given the functions in it, which will be focused on calculating probabilities to choose the best words. A game.py module is added and it will contain functions directly related to running the game. Finally, wordle_solver.py is added to the repo, it is the file that will be run to play the game. So far, it contains mostly an outline and is a work in progress.