fixed an error in check_word
Fixed the check_word function to account for the following specific case:
If a guess has twice the same letter, such as 'again', and one of the two 'a's is in the solution, such as 'axiom', then the previous version of check_word would delete the option 'axiom' because it has an 'a' and 'a' is both in the green_letters and in the black_letters. Now the function is coded in a way to make sure that this doesn't happen. A test has also been added to check this.