Added count_difference functions
Added a function in words.py that compares 2 words and outputs the number of letters that are different. This was then incorportated in probabilities.py in a function that compares a whole list of words to a guess. The goal is to use these for the second guess so that we get a maximum amount of information from it (by choosing a second guess that is very different from the 1st)