Added randomized selection of words
In each different case of suggesting a word, instead of picking the one with the highest score, it uses the scores as a probability distribution (after renormalizing them) and then picks randomly from that distribution. In guess 2, it choses a word with a maximum amount of different letters from guess1 but it also considers the frequency in the language.